refacto: Restructure Docker build process and update supporting files
The build process for creating the Docker image for Prestashop has been significantly restructured. This change was due to the necessity to support different PHP versions, OS flavours, and server types, which all affected the creation of the Docker images. Thus, build.sh was updated to get PHP version and flavour based on the Prestashop version and considered OS and server types. Also, all the json files that were used to get info for build.sh and were not relevant anymore were removed.

In addition, the Github workflows have been updated to work with the new build process structure. Also, Dockerfile names have been simplified removing '-base' from the names.